Commit to phpBB4 and PHP 8.2
