+++ .github/workflows/tests.yml
@@ -4,9 +4,11 @@ on:
     push:
         branches:
             - master
+            - epv-phpbb4
     pull_request:
         branches:
             - master
+            - epv-phpbb4
 
 jobs:
     php-tests:
@@ -14,14 +16,10 @@ jobs:
         strategy:
             matrix:
                 include:
-                    - php: '7.2'
-                    - php: '7.3'
-                    - php: '7.4'
-                    - php: '8.0'
-                    - php: '8.1'
                     - php: '8.2'
                     - php: '8.3'
                     - php: '8.4'
+                    - php: '8.5'
 
         name: PHP ${{ matrix.php }}
 
@@ -47,9 +45,6 @@ jobs:
               with:
                   path: /tmp/composer-cache
                   key: ${{ runner.os }}-${{ hashFiles('**/composer.lock') }}
-
-            - name: Delete composer.lock (library mode)
-              run: rm -f composer.lock
 
             - name: Install Composer dependencies
               uses: php-actions/composer@v6
