+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "phpbb/epv",
-    "description": "A extension validator for phpBB extensions. Extensions are required to pass the validator when submitted to the extension database.",
+    "description": "An extension validator for phpBB extensions. Extensions are required to pass the validator when submitted to the extension database.",
     "license": "GPL-2.0-only",
     "authors": [
         {
@@ -10,19 +10,19 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=7.2",
+        "php": ">=8.2",
         "ext-json": "*",
-        "symfony/yaml": "^3.0 || ^4.0 || ^5.0 || ^6.0 || ^7.0",
-        "symfony/console": "^3.0 || ^4.0 || ^5.0 || ^6.0 || ^7.0",
-        "symfony/finder": "^3.0 || ^4.0 || ^5.0 || ^6.0 || ^7.0",
-        "symfony/process": "^3.0 || ^4.0 || ^5.0 || ^6.0 || ^7.0",
+        "symfony/yaml": "^7.4",
+        "symfony/console": "^7.4",
+        "symfony/finder": "^7.4",
+        "symfony/process": "^7.4",
         "nikic/php-parser": "^4.0 || ^5.0",
         "gitonomy/gitlib": "^1.3.0",
         "sensiolabs/ansi-to-html": "~1.1",
-        "composer/composer": "^1.5 || ^2.0"
+        "composer/composer": "^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.5 || ^7.0 || ^8.0 || ^9.0"
+        "phpunit/phpunit": "^9.0 || ^10.0"
     },
     "bin": [
         "src/EPV.php"
