+++ phpunit.xml.dist
@@ -1,20 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
-<phpunit backupGlobals="true"
-         backupStaticAttributes="true"
+<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         backupGlobals="true"
          colors="true"
-         convertErrorsToExceptions="true"
-         convertNoticesToExceptions="true"
-         convertWarningsToExceptions="true"
          processIsolation="false"
          stopOnFailure="false"
          bootstrap="tests/bootstrap.php"
-        >
-    <testsuites>
-        <testsuite name="phpBB EPV Test Suite">
-            <directory suffix="_test.php">./tests/</directory>
-
-        </testsuite>
-    </testsuites>
-
+         xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/10.5/phpunit.xsd"
+         cacheDirectory=".phpunit.cache"
+         backupStaticProperties="true">
+  <testsuites>
+    <testsuite name="phpBB EPV Test Suite">
+      <directory suffix="_test.php">./tests/</directory>
+    </testsuite>
+  </testsuites>
 </phpunit>
